pyproject.toml
[project]
name = "realmate-challenge"
version = "0.1.0"
description = ""
authors = [{ name = "Seu nome aqui", email = "seu.email@email.com.br" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["django (>=5.1.6,<6.0.0)", "djangorestframework (>=3.16.0,<4.0.0)", "python-dateutil (>=2.9.0.post0,<3.0.0)", "channels (>=4.2.2,<5.0.0)", "channels-redis (>=4.2.1,<5.0.0)", "daphne (>=4.1.2,<5.0.0)", "django-redis (>=5.4.0,<6.0.0)", "redis (>=6.0.0,<7.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.1"

